requirements.txt
streamlit
supabase
pandas
folium
streamlit-folium
geopy
requests
python-dotenv
fpdf2
